+++ src/main/java/com/grepp/moodlink/app/model/admin/book/repository/AdminBookGenreRepository.java
@@ -1,4 +1,4 @@
-package com.grepp.moodlink.app.model.admin.book;
+package com.grepp.moodlink.app.model.admin.book.repository;
 
 import com.grepp.moodlink.app.model.data.book.entity.BookGenre;
 import java.util.List;
