+++ src/main/java/com/grepp/moodlink/app/model/admin/book/repository/AdminBookRepository.java
@@ -1,4 +1,4 @@
-package com.grepp.moodlink.app.model.admin.book;
+package com.grepp.moodlink.app.model.admin.book.repository;
 
 import com.grepp.moodlink.app.model.data.book.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
