+++ src/main/java/com/grepp/moodlink/app/model/admin/book/repository/AdminBookRepositoryCustom.java
@@ -1,6 +1,5 @@
-package com.grepp.moodlink.app.model.admin.book;
+package com.grepp.moodlink.app.model.admin.book.repository;
 
-import com.grepp.moodlink.app.model.data.book.dto.BookDto;
 import com.grepp.moodlink.app.model.data.book.entity.Book;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
