+++ src/main/java/com/grepp/moodlink/app/model/admin/book/repository/AdminBookRepositoryImpl.java
@@ -1,6 +1,5 @@
-package com.grepp.moodlink.app.model.admin.book;
+package com.grepp.moodlink.app.model.admin.book.repository;
 
-import com.grepp.moodlink.app.model.data.book.dto.BookDto;
 import com.grepp.moodlink.app.model.data.book.entity.Book;
 import com.grepp.moodlink.app.model.data.book.entity.QBook;
 import com.querydsl.jpa.impl.JPAQuery;
@@ -14,7 +13,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Repository
